+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # chrona-core
 
+## 0.3.5
+
 ## 0.3.4
 
 ### Patch Changes